change task05

+++ Lesson04/Task05/Reverser.java
@@ -1,7 +1,7 @@
 public class Reverser {
     public static void main(String[] args) {
-        long number = 3456789;
-        long reverseNumber = ReverserLogic.reversOrder(number);
+        int number = 3456789;
+        int reverseNumber = ReverserLogic.reversOrder(number);
         Printer.print(reverseNumber, number);
     }
 }
